Split Residencies collection resources
