+++ app/collections/residencies/index.js
@@ -1,109 +1,5 @@
-import SimpleSchema from 'simpl-schema';
-import UserEventLog from '/collections/user_event_log/';
+import ResidenciesSchema from './schema';
 
 export default Residencies = new Mongo.Collection('residencies');
 
-var ResidenciesSchema = new SimpleSchema({
-  residentId: {
-    type: String,
-  },
-  homeId: {
-    type: String,
-  },
-  moveIn: {
-    type: Date,
-    autoValue: function () {
-      if (this.isSet) {
-        // Get move in date from form
-        const moveInDate = new Date(this.value);
-
-        // set move in to midnight UTC
-        moveInDate.setUTCHours(0,0,0,0);
-
-        return moveInDate;
-      }
-    },
-  },
-  moveOut: {
-    type: Date,
-    optional: true,
-    autoValue: function () {
-      if (this.isSet) {
-        // Get move out date from form
-        const moveOutDate = new Date(this.value);
-
-        if (moveOutDate) {
-          // set move out to midnight UTC
-          moveOutDate.setUTCHours(0,0,0,0);
-
-          return moveOutDate;
-        }
-      }
-    },
-  },
-});
-
 Residencies.attachSchema(ResidenciesSchema);
-
-Residencies.allow({
-  'insert': function () {
-    // Get user ID
-    let userId = Meteor.userId();
-
-    // Placeholder for administrator check
-    let userIsAdministrator;
-
-    // Placeholder for insert privilege check
-    let userCanInsert;
-
-    if (userId) {
-      // Check if user is administrator
-      userIsAdministrator = Roles.userIsInRole(userId, ['admin']);
-    }
-
-    // Only allow adminstator users insert
-    userCanInsert = (userId && userIsAdministrator);
-
-    return userCanInsert;
-  },
-  'update': function () {
-    // Get user ID
-    let userId = Meteor.userId();
-
-    // Placeholder for administrator check
-    let userIsAdministrator;
-
-    // Placeholder for update privilege check
-    let userCanUpdate;
-
-    if (userId) {
-      // Check if user is administrator
-      userIsAdministrator = Roles.userIsInRole(userId, ['admin']);
-    }
-
-    // Only allow adminstator users insert
-    userCanUpdate = (userId && userIsAdministrator);
-
-    return userCanUpdate;
-  },
-  'remove': function () {
-    // Get user ID
-    let userId = Meteor.userId();
-
-    // Placeholder for administrator check
-    let userIsAdministrator;
-
-    // Placeholder for insert privilege check
-    let userCanRemove;
-
-    if (userId) {
-      // Check if user is administrator
-      userIsAdministrator = Roles.userIsInRole(userId, ['admin']);
-    }
-
-    // Only allow adminstator users insert
-    userCanRemove = (userId && userIsAdministrator);
-
-    return userCanRemove;
-  }
-});
